feat(MAM-28): 새로운 맛집 카드 컴포넌트 추가 및 기존 UI 개선
- 기존 맛집 리스트를 새로운 NewRestaurantCard 컴포넌트로 대체하여 코드 구조를 개선함.
- NewRestaurantCard 컴포넌트에서 맛집 정보를 표시하는 UI를 새롭게 구성하여 사용자 경험을 향상시킴.
